requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
selenium==4.15.2
webdriver-manager==4.0.1 